fix: narrow terminal test PR scope
